Feat: Renamed Following component to SearchFriends and made a new Following component that renders the users I follow
